requirements.txt
# Core AI (GPU Native)
llama-cpp-python==0.2.90
torch==2.1.2+cu121
torchvision==0.16.2+cu121
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors==0.4.2
ip-adapter==0.1.0

# Utilities
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
boto3==1.34.0

# YouTube API
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
